Plasma!
I got ChatGPT to help with boids, which was helpful even though I still
had to tune it by hand (this version is much nicer), but it fell flat on
its face with plasma.  I basically had to write the whole thing by hand
(based on The Demo Effects Collection).

Also made minor changes to utils as well as renamed Renderer draw() to
frame().  Also properly cleaning up when switching sketches, so the
memory slop should be fixed.